+++ examples/learnopengl/1_getting_started/6_1_coordinate_systems.ex
@@ -208,13 +208,11 @@ defmodule EAGL.Examples.LearnOpenGL.GettingStarted.CoordinateSystems do
 
     # Compile and link shaders
     with {:ok, vertex_shader} <-
-           create_shader(
-             @gl_vertex_shader,
+           create_shader(:vertex,
              "learnopengl/1_getting_started/6_1_coordinate_systems/vertex_shader.glsl"
            ),
          {:ok, fragment_shader} <-
-           create_shader(
-             @gl_fragment_shader,
+           create_shader(:fragment,
              "learnopengl/1_getting_started/6_1_coordinate_systems/fragment_shader.glsl"
            ),
          {:ok, program} <- create_attach_link([vertex_shader, fragment_shader]) do