+++ examples/learnopengl/1_getting_started/6_4_coordinate_systems_exercise.ex
@@ -178,13 +178,11 @@ defmodule EAGL.Examples.LearnOpenGL.GettingStarted.CoordinateSystemsExercise do
 
     # Compile and link shaders
     with {:ok, vertex_shader} <-
-           create_shader(
-             @gl_vertex_shader,
+           create_shader(:vertex,
              "learnopengl/1_getting_started/6_4_coordinate_systems_exercise/vertex_shader.glsl"
            ),
          {:ok, fragment_shader} <-
-           create_shader(
-             @gl_fragment_shader,
+           create_shader(:fragment,
              "learnopengl/1_getting_started/6_4_coordinate_systems_exercise/fragment_shader.glsl"
            ),
          {:ok, program} <- create_attach_link([vertex_shader, fragment_shader]) do