+++ examples/learnopengl/1_getting_started/7_5_camera_exercise1.ex
@@ -125,14 +125,12 @@ defmodule EAGL.Examples.LearnOpenGL.GettingStarted.CameraExercise1 do
 
     # Compile shaders
     {:ok, vertex} =
-      create_shader(
-        @gl_vertex_shader,
+      create_shader(:vertex,
         "learnopengl/1_getting_started/7_5_camera_exercise1/vertex_shader.glsl"
       )
 
     {:ok, fragment} =
-      create_shader(
-        @gl_fragment_shader,
+      create_shader(:fragment,
         "learnopengl/1_getting_started/7_5_camera_exercise1/fragment_shader.glsl"
       )