+++ examples/learnopengl/2_lighting/1_1_colors.ex
@@ -175,13 +175,11 @@ defmodule EAGL.Examples.LearnOpenGL.Lighting.Colors do
 
     # Compile and link lighting shader (for the object cube)
     with {:ok, lighting_vertex_shader} <-
-           create_shader(
-             @gl_vertex_shader,
+           create_shader(:vertex,
              "learnopengl/2_lighting/1_1_colors/lighting_vertex_shader.glsl"
            ),
          {:ok, lighting_fragment_shader} <-
-           create_shader(
-             @gl_fragment_shader,
+           create_shader(:fragment,
              "learnopengl/2_lighting/1_1_colors/lighting_fragment_shader.glsl"
            ),
          {:ok, lighting_program} <-
@@ -190,14 +188,12 @@ defmodule EAGL.Examples.LearnOpenGL.Lighting.Colors do
 
       # Compile and link light cube shader (for the light source cube)
       {:ok, light_cube_vertex_shader} =
-        create_shader(
-          @gl_vertex_shader,
+        create_shader(:vertex,
           "learnopengl/2_lighting/1_1_colors/light_cube_vertex_shader.glsl"
         )
 
       {:ok, light_cube_fragment_shader} =
-        create_shader(
-          @gl_fragment_shader,
+        create_shader(:fragment,
           "learnopengl/2_lighting/1_1_colors/light_cube_fragment_shader.glsl"
         )