+++ examples/learnopengl/2_lighting/2_2_basic_lighting_specular.ex
@@ -223,13 +223,11 @@ defmodule EAGL.Examples.LearnOpenGL.Lighting.BasicLightingSpecular do
 
     # Compile and link lighting shader (for the object cube with specular lighting)
     with {:ok, lighting_vertex_shader} <-
-           create_shader(
-             @gl_vertex_shader,
+           create_shader(:vertex,
              "learnopengl/2_lighting/2_2_basic_lighting_specular/lighting_vertex_shader.glsl"
            ),
          {:ok, lighting_fragment_shader} <-
-           create_shader(
-             @gl_fragment_shader,
+           create_shader(:fragment,
              "learnopengl/2_lighting/2_2_basic_lighting_specular/lighting_fragment_shader.glsl"
            ),
          {:ok, lighting_program} <-
@@ -238,14 +236,12 @@ defmodule EAGL.Examples.LearnOpenGL.Lighting.BasicLightingSpecular do
 
       # Compile and link light cube shader (for the light source cube)
       {:ok, light_cube_vertex_shader} =
-        create_shader(
-          @gl_vertex_shader,
+        create_shader(:vertex,
           "learnopengl/2_lighting/2_2_basic_lighting_specular/light_cube_vertex_shader.glsl"
         )
 
       {:ok, light_cube_fragment_shader} =
-        create_shader(
-          @gl_fragment_shader,
+        create_shader(:fragment,
           "learnopengl/2_lighting/2_2_basic_lighting_specular/light_cube_fragment_shader.glsl"
         )