+++ examples/learnopengl/2_lighting/2_4_basic_lighting_exercise2.ex
@@ -238,13 +238,11 @@ defmodule EAGL.Examples.LearnOpenGL.Lighting.BasicLightingExercise2 do
 
     # Compile and link lighting shader (view space version)
     with {:ok, lighting_vertex_shader} <-
-           create_shader(
-             @gl_vertex_shader,
+           create_shader(:vertex,
              "learnopengl/2_lighting/2_4_basic_lighting_exercise2/lighting_vertex_shader.glsl"
            ),
          {:ok, lighting_fragment_shader} <-
-           create_shader(
-             @gl_fragment_shader,
+           create_shader(:fragment,
              "learnopengl/2_lighting/2_4_basic_lighting_exercise2/lighting_fragment_shader.glsl"
            ),
          {:ok, lighting_program} <-
@@ -255,14 +253,12 @@ defmodule EAGL.Examples.LearnOpenGL.Lighting.BasicLightingExercise2 do
 
       # Compile and link light cube shader
       {:ok, light_cube_vertex_shader} =
-        create_shader(
-          @gl_vertex_shader,
+        create_shader(:vertex,
           "learnopengl/2_lighting/2_4_basic_lighting_exercise2/light_cube_vertex_shader.glsl"
         )
 
       {:ok, light_cube_fragment_shader} =
-        create_shader(
-          @gl_fragment_shader,
+        create_shader(:fragment,
           "learnopengl/2_lighting/2_4_basic_lighting_exercise2/light_cube_fragment_shader.glsl"
         )